Update main.js
Added better functionality and tried to prevent circles from overlapping, but they still can-- my guess is that as one circle collides into another, it pushes that second circle into a third circle which has already been looped through, meaning that a collision doesn't happen. I tried to implement a chain reaction function for this, but if a circle is lodged in between a lot of circles at once, it will continuously collide back and forth, leading to a call stack error. Also, I need to check the collision equations because the velocities spin out of control (this could also be due to the overlapping issue though).